+++ src/test/groovy/com/cedarsoftware/ncube/TestNCubeJdbcPersister.groovy
@@ -37,6 +37,53 @@ class TestNCubeJdbcPersister extends NCubeCleanupBaseTest
     private ApplicationID defaultSnapshotApp = new ApplicationID(ApplicationID.DEFAULT_TENANT, APP_ID, "1.0.0", ApplicationID.DEFAULT_STATUS, ApplicationID.TEST_BRANCH)
 
     @Test
+    void testCopyBranchInitialRevisions()
+    {
+        String cubeNameBranch = 'TestBranch'
+        String cubeNameAge = 'TestAge'
+        ApplicationID branch1 = defaultSnapshotApp.asBranch('branch1')
+        ApplicationID branch2 = defaultSnapshotApp.asBranch('branch2')
+        ApplicationID branch3 = defaultSnapshotApp.asBranch('branch3')
+        preloadCubes(branch1, 'test.branch.1.json', 'test.branch.age.1.json')
+        mutableClient.commitBranch(branch1)
+        mutableClient.updateBranch(branch2)
+
+        NCube ncube1 = mutableClient.getCube(branch1, cubeNameBranch)
+        ncube1.setCell('XYZ', [Code: -10])
+        mutableClient.updateCube(ncube1)
+        mutableClient.commitBranch(branch1)
+
+        NCube ncube2 = mutableClient.getCube(branch2, cubeNameBranch)
+        ncube2.setCell('XYZ', [Code: 10])
+        mutableClient.updateCube(ncube2)
+        // make sure this works with multiple modified cubes in branch
+        ncube2 = mutableClient.getCube(branch2, 'TestAge')
+        ncube2.setCell('infant', [Code: 0])
+        mutableClient.updateCube(ncube2)
+
+        ncube1.setCell('ABC', [Code: -10])
+        mutableClient.updateCube(ncube1)
+        mutableClient.commitBranch(branch1)
+
+        mutableClient.copyBranch(branch2, branch3)
+
+        NCube ncube3 = mutableClient.getCube(branch3, cubeNameBranch)
+        assert 'ABC' == ncube3.getCell([Code: -10])
+        assert 'XYZ' == ncube3.getCell([Code: 10])
+
+        ncube3 = mutableClient.getCube(branch3, cubeNameAge)
+        assert 'infant' == ncube3.getCell([Code: 0])
+
+        mutableClient.rollbackBranch(branch3, [cubeNameBranch, cubeNameAge] as Object[])
+        ncube3 = mutableClient.getCube(branch3, cubeNameBranch)
+        assert 'ABC' == ncube3.getCell([Code: -10])
+        assert 'GHI' == ncube3.getCell([Code: 10])
+
+        ncube3 = mutableClient.getCube(branch3, cubeNameAge)
+        assert 'baby' == ncube3.getCell([Code: 0])
+    }
+
+    @Test
     void testDbApis()
     {
         NCube ncube1 = NCubeBuilder.testNCube3D_Boolean